+++ custom.d.ts
@@ -1,8 +1,10 @@
-declare module '*.scss';
+declare module '*.scss'
 
 declare module '*.svg' {
-  import React from 'react';
-  const content: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
-  export default content;
-  export const ReactComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
+  import React from 'react'
+  const content: React.FunctionComponent<React.SVGProps<SVGSVGElement>>
+  export default content
+  export const ReactComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement>>
 }
+
+declare module '*.png'