+++ src/PhoneInput/PhoneInput.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode, useState } from 'react'
 
-import clsx from 'clsx'
+import { clsx } from 'clsx'
 
 import styles from './PhoneInput.module.scss'