+++ src/Table/Table.tsx
@@ -1,9 +1,9 @@
 import React, { ReactNode } from 'react'
 
-import clsx from 'clsx'
+import { clsx } from 'clsx'
 
 import styles from './Table.module.scss'
-import DataIcon from '../../icons/system/data.svg'
+import DataIcon from '../../assets/icons/system/data.svg'
 
 interface ColumnProps<T> {
   key: string