+++ src/Textarea/Textarea.tsx
@@ -1,6 +1,6 @@
 import React, { TextareaHTMLAttributes } from 'react'
 
-import clsx from 'clsx'
+import { clsx } from 'clsx'
 
 import styles from './Textarea.module.scss'