+++ src/Toast/Toast.tsx
@@ -1,9 +1,9 @@
 import { ReactNode, useEffect } from 'react'
 
-import clsx from 'clsx'
+import { clsx } from 'clsx'
 
 import styles from './Toast.module.scss'
-import CloseIcon from '../../icons/system/close.svg'
+import CloseIcon from '../../assets/icons/system/close.svg'
 
 import Typography from '@/Typography'